requirements.txt
django
pymysql
firebase-admin
python-decouple
djangorestframework
django-cors-headers
sendgrid
channels
channels-redis
daphne